Expanded regex pattern matching.
This is a fix to an problem with the program. It would needlessly
exclude certain files based on their names as they failed to match a
regex pattern. This commit expands the pattern to capture more artist,
album and song names.

A test file has been included for a set of file names that previously
failed to be pattern matched.