Bug Fix and Error Checking.
A bug was found in the bc_renamer.py script. If a songname contained a
' - ' character the script would fail. This was fixed, naievly, to split
the filename on a maximum of 2 ' - ' chars. Naturally, if the artist or
album name contains such characters, then it would still work, but would
end up putting the songs in a nonsensical directory.

A bunch of error checking was added to the shell script to ensure that
the required commands are installed, and that the specified directories
exist.